Fix python wrappers for std::array and sdt::bitset used by streamtrace Add support for python callbacks to be used as function arguments to trace::scan and trace::filter
The trace::scan and trace::filter methods are exported as trace::scan_trace and trace::filter_trace
in the python interface. These accept a proxy python object deriving from the Scanner, DetailedScanner or
Filter classes provided as builtins by the cheritrace bindings. These objects act as a proxy to wrap
the callback to the scan() and filter() methods, e.g. Scanner.run is invoked as the callback during a call
to trace::scan.